Ventana Stats, muestra los días. Las imágenes 100% se van a cargar y …

+++ app/src/main/java/com/example/easysoccer1/ui/view/DescriptionSportCenterActivity.kt
@@ -94,7 +94,7 @@ class DescriptionSportCenterActivity : AppCompatActivity() {
                 nameUser = user.getOrNull()!!.nameUser,
                 comment = binding.textComment.text.toString(),
                 nameSportCenter = sportCenter.getOrNull()!!.nameSportCenter,
-                id = idComment.toString().toInt()
+                id = idComment.toString()
             )
         )
